backend/requirements.txt
# AMAR MVP Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# LangGraph and LLM
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=0.1.0
openai>=1.0.0
groq>=0.4.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1

# HTTP Client for deployment APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0

# File handling and utilities
aiofiles==23.2.1

# System monitoring
psutil==5.9.6

# RAG and Vector Search Dependencies
sentence-transformers==3.0.0
chromadb==0.5.0
numpy==1.26.4
scikit-learn==1.5.0